add the ability to turn a USB port on and off
LightBox currently has very little to do with actual lights. The code
manages the on/off state of a designated USB port and shows that on/off
status in an icon added to the Apple menubar. Clicking the icon will
also toggle the on/off state. Super janky tray app!

Also kinda revamped how the zoom detection works:

* changed the in-meeting detection to look in the Zoom app's pull down
  menus; if "Meeting" in present as a pull-down, you're in a meeting.
  Needed to do this because the window title detection would miss when I
  was screen sharing.
* used a Hammerspoon application watcher to check for starting/stopping
  and focus/unfocus of the Zoom application to decide whether to run the
  timer that checks for in-meeting status. The timer now stops polling
  for whether you're in a meeting if Zoom isn't running.

Signed-off-by: Robb Kidd <robb@thekidds.org>