add a guard for when Zoom does not yet have menus
It is possible for Zoom to start and run without menu items.

    # Zoom running ...
    > zoom_app = hs.application.find("zoom.us")
    > zoom_app
    hs.application: zoom.us (0x60800005d1b8)
    
    # ... but no pull-down menus
    > menu_items = zoom_app:getMenuItems()
    > menu_items
    nil

This can happen, for example, after a force-reboot of a Mac where the
user has opted to reopen the applications that were running before
reboot. Zoom appears to be in the application list, but in what seems to
be a partially initialized state.