Change method of detecting being in a zoom meeting
Zoom recently updated and the main zoom window has the title of "Zoom",
which is the same as the in meeting zoom window when the toolbar
disappears. This method now checks for a "Start meeting" menu item (it's
in the "zoom.us" application menu) and checks to see if it's enabled or
not. If it's disabled, then you're in a meeting. If it's enabled, then
you're not in a meeting. If it's missing, then assume you're not in a
meeting (this will probably happen in the case of a new zoom update that
changes the menu).