Add expiration feature
Slack's `users.profile.set` API allows a profile parameter `status_expiration` which takes an integer value: either seconds until expiration, or, 0 to mean never expires. I added this functionality because I wanted to be able to set expiration for status like "lunch" to an hour. Also, the sample's "zoom" profile does not work unless you have `:zoom:` emoji installed in your Slack, so I changed it to `ghost` which is a standard emoji that Slack has be default. Hope this helps others.